+++ ichnaea/api/locate/geoip.py
@@ -28,10 +28,15 @@ class GeoIPSource(Source):
         # there's no need to do a lookup again.
         geoip = query.geoip
         if geoip:
+            fallback = self.fallback_field
+            if query.geoip_only:
+                # Don't set a fallback if only IP data was provided.
+                fallback = None
             results.add(self.result_type(
                 lat=geoip['latitude'],
                 lon=geoip['longitude'],
                 accuracy=geoip[self.geoip_accuracy_field],
+                fallback=fallback,
                 region_code=geoip['region_code'],
                 region_name=geoip['region_name'],
                 score=geoip['score'],