+++ ichnaea/api/locate/tests/test_locate_v0.py
@@ -56,7 +56,6 @@ class LocateV0Base(BaseLocateTest):
             'lat': london['latitude'],
             'lon': london['longitude'],
             'accuracy': london['radius'],
-            'fallback': 'ipf',
         }
 
     def check_model_response(self, response, model,
@@ -281,15 +280,4 @@ class TestView(LocateV0Base, CommonLocateTest, CommonPositionTest):
 
 
 class TestError(LocateV0Base, CommonLocateErrorTest):
-
-    def test_apikey_error(self, app, data_queues,
-                          db_rw_drop_table, raven, ro_session, stats):
-        super(TestError, self).test_apikey_error(
-            app, data_queues, db_rw_drop_table,
-            raven, ro_session, stats, db_errors=1)
-
-    def test_database_error(self, app, data_queues,
-                            db_rw_drop_table, raven, ro_session, stats):
-        super(TestError, self).test_database_error(
-            app, data_queues, db_rw_drop_table,
-            raven, ro_session, stats, db_errors=5)
+    pass